changed calculations of effective prec and ETa_rf
Effective precipitation is now calculated for the whole year, not just the growing period.
ETa_rf for the first day is set = ETc, meaning that on day==0 we assume maximum evapotranspiration potential. Alternative is setting ETa_rf=0, but it should not make a big difference as the value is very low for Jan 1st anyway.